Increased tf prediction time precision

+++ Demo/src/TFPredictor.h
@@ -15,7 +15,7 @@ public:
 	bool predict(void* data, unsigned int dataSize);
 
 	float getLastPredictionValue() const;
-	long long getLastPredictionTime() const;
+	double getLastPredictionTime() const;
 
 private:
 	TF_Tensor* createTensor(void* inputData, unsigned int dataSize, TensorShape* inputShape);
@@ -24,6 +24,6 @@ private:
 	std::unique_ptr<MySession> m_session;
 	TensorShape* m_inputShape;
 	float m_lastPrediction;
-	long long m_lastPredictionTime;
+	double m_lastPredictionTime;
 
 };
